requirements.txt
botbuilder-integration-aiohttp==4.16.2
aiohttp>=3.8.1
aiohttp_cors 
python-dotenv
botbuilder-core==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
msrest
requests


# Azure SDKs for RAG Runtime
azure-identity        # For authentication
azure-search-documents # For Azure AI Search (use stable release)
azure-storage-blob # For Blob Storage interaction
openai      # For Azure OpenAI / OpenAI (v1.x required)


# RAG Utilities
tiktoken            # For token counting
tenacity
msal
cryptography
PyJWT
Pillow
types-Pillow
pypdf
PyMuPDF
beautifulsoup4
types-beautifulsoup4
msgraph-sdk
python-dotenv
prompty==0.1.50
rich
typing-extensions
PyYAML